==> On Dataset 1/Phase 1 - Creating Dataset/00-zerosteps.ipynb <==
import pandas as pd

textFileName = "rockyou.txt"
csvFileName = "rockyou.csv"

with open(textFileName, 'r', encoding='latin-1') as file:
    lines = file.readlines()
    
lines = [line.strip() for line in lines]    # Remove any extra whitespace (e.g., newline characters)
dataframe = pd.DataFrame(lines, columns=["CommonPasswords"])
dataframe.to_csv(csvFileName, index=False)
print(f"CSV file saved at: {csvFileName}")
