==> On Dataset 1/Phase 1 - Creating Dataset/01-preprocess.ipynb <==
<h1>1. Necessary Imports & Steps</h1>

import pandas as pd
import numpy as np
import re

<h1>2. Read the Original <b>PWNED</b> Dataset</h1>
<ul>
    <li>Password: Representing history of passwords been compromised before</li>
    <li>Hash: Representing their corresponding hashes</li>
    <li>Count: Representing how many times have they been reported compromised</li>
</ul>

# Read target dataset
dataset = pd.read_csv("pwned.csv")
dataset

# Removing hash from target dataset - not necessary for us at this moment
dataset.drop(columns=["hash"], inplace=True)
dataset

# Add new feature 'length' to target dataset - count the number of letters in the password
dataset['length'] = dataset['password'].str.len()
dataset['password'] = dataset['password'].fillna("").astype(str)
dataset

# Read common password's dataset - rockyou has millions of common passwords in the history
commonPasswords = pd.read_csv("rockyou.csv", header=None)
commonPasswords

<h1>3. Preprocessing Steps - Getting the Dataset Ready</h1>

# If a password in PWNED dataset is common/1 or not/0 - comparing each password from PWNED to all the password of ROCKYOU
commonSet = set(commonPasswords.iloc[:, 0]) # Convert commonPasswords to a set (exclude repetitive values)
dataset['common'] = np.vectorize(lambda x: 1 if x in commonSet else 0)(dataset['password']) # Use a NumPy vectorized function for speed
dataset

def hasAlphabets(password):
    return 1 if any(char.isalpha() for char in password) else 0
def hasNumbers(password):
    return 1 if any(char.isdigit() for char in password) else 0
def hasSpecialChars(password):
    return 1 if any(not char.isalnum() for char in password) else 0

dataset['alphabets'] = dataset['password'].apply(hasAlphabets)
dataset['numbers'] = dataset['password'].apply(hasNumbers)
dataset['special'] = dataset['password'].apply(hasSpecialChars)
dataset

def countAlphabets(password):
    return sum(char.isalpha() for char in password)
def countNumbers(password):
    return sum(char.isdigit() for char in password)
def countSpecialChars(password):
    return sum(not char.isalnum() for char in password)

dataset['alpha-count'] = dataset['password'].apply(countAlphabets)
dataset['numeric-count'] = dataset['password'].apply(countNumbers)
dataset['special-count'] = dataset['password'].apply(countSpecialChars)
dataset

def hasUppercase(password):
    return 1 if any(char.isupper() for char in password) else 0
def hasLowercase(password):
    return 1 if any(char.islower() for char in password) else 0

dataset['uppercase'] = dataset['password'].apply(hasUppercase)
dataset['lowercase'] = dataset['password'].apply(hasLowercase)
dataset



def hasSequentialPatterns(password):
    alphabeticSequence = 'abcdefghijklmnopqrstuvwxyz'
    numericSequence = '0123456789'
    commonPatterns = ['qwe', 'asd', 'zxc', 'ghj', 'iop', 'jkl', 'bnm']

    for pattern in commonPatterns:
        if pattern in password.lower():
            return 1
    if any(password.lower() == alphabeticSequence[i:i+len(password)] for i in range(len(alphabeticSequence)-len(password)+1)):
        return 1
    if any(password == numericSequence[i:i+len(password)] for i in range(len(numericSequence)-len(password)+1)):
        return 1
    return 0

dataset['sequential-pattern'] = dataset['password'].apply(hasSequentialPatterns)
dataset

def hasRepeatedCharacters(password):
    # Check if the password contains a sequence of the same character repeated, e.g., 'aaaa', 'bbbb', '@@@@', etc.
    if re.search(r"(.)\1{2,}", password):  # Checks for any character repeated 3 or more times consecutively
        return 1
    return 0

dataset['repeated-characters'] = dataset['password'].apply(hasRepeatedCharacters)
dataset


def caseRatio(password):
    # Count lowercase and uppercase characters
    lowercase_count = sum(1 for char in password if char.islower())
    uppercase_count = sum(1 for char in password if char.isupper())
    
    # Avoid division by zero if no uppercase letters
    if uppercase_count == 0:
        return 0
    else:
        return lowercase_count / uppercase_count

dataset['case-ratio'] = dataset['password'].apply(caseRatio)
dataset

dataset.drop('alphabets', axis=1, inplace=True)
dataset.drop('numbers', axis=1, inplace=True)
dataset.drop('special', axis=1, inplace=True)
dataset

<h1>4. Saving the Final Dataset (No Labels Yet)</h1>

dataset.to_csv("Featured-Dataset.csv", index=False)

***
