==> On Dataset 1/Phase 1 - Creating Dataset/02-2-guassianmixture.ipynb <==
import pandas as pd
dataset = pd.read_csv("Featured-Dataset.csv")
dataset = dataset.dropna()
dataset = dataset.drop(columns=['password'])
dataset

# Guassian Mixture
from sklearn.mixture import GaussianMixture
from sklearn.preprocessing import StandardScaler

# Scale the data
scaler = StandardScaler()
scaledFeatures = scaler.fit_transform(dataset[['count', 'length', 'common', 'alpha-count', 'numeric-count', 
                                                'special-count', 'uppercase', 'lowercase', 'sequential-pattern', 
                                                'repeated-characters', 'case-ratio']])

# Apply GMM with 3 components (clusters)
GMM = GaussianMixture(n_components=3, random_state=42)
gmmLabels = GMM.fit_predict(scaledFeatures)

labelMapping = {0: 'weak', 1: 'moderate', 2: 'strong'}
dataset['strength'] = pd.Series(gmmLabels).map(labelMapping)


strong_count = dataset[dataset['strength'] == 'strong'].shape[0]
print(f"Number of strong passwords after relaxing rules: {strong_count}")

moderate_count = dataset[dataset['strength'] == 'moderate'].shape[0]
print(f"Number of moderate passwords after relaxing rules: {moderate_count}")

weak_count = dataset[dataset['strength'] == 'weak'].shape[0]
print(f"Number of weak passwords after relaxing rules: {weak_count}")
