==> On Dataset 1/Phase 1 - Creating Dataset/03-kmeanslabelling.ipynb <==
<h1>1. Necessary Imports & Steps</h1>

import pandas as pd
from sklearn.cluster import KMeans
from sklearn.preprocessing import StandardScaler

<h1>2. Read Featured-Dataset</h1>

# Read the preprocessed dataset and drop NaN values
dataset = pd.read_csv("Featured-Dataset.csv")
dataset = dataset.dropna()
dataset = dataset.drop(columns=['password'])
dataset

<h1>3. Apply K-Means with 10 Clusters - Based on Elbow Method (<small><i>03-howmanyclusters.ipynb</i></small>)</h1>

# Scale the data first (important for K-Means)
scaler = StandardScaler()
scaledFeatures = scaler.fit_transform(dataset[['count', 'length', 'common', 'alpha-count', 'numeric-count', 
                                                'special-count', 'uppercase', 'lowercase', 'sequential-pattern', 
                                                'repeated-characters', 'case-ratio']])

KMeans = KMeans(n_clusters=10, random_state=42)
dataset['strength'] = KMeans.fit_predict(scaledFeatures)
dataset

<h1>4. Labelling First 3 Classes as 'Weak', Next 4 as 'Moderate' and Last 3 as 'Strong'</h1>

# Map the numerical strength to categories
dataset['strength'] = dataset['strength'].replace({0: 'weak', 1: 'weak', 2: 'weak', 
                                                   3: 'moderate', 4: 'moderate', 5: 'moderate', 6: 'moderate', 
                                                   7: 'strong', 8: 'strong', 9: 'strong'})
dataset

strongCount = dataset[dataset['strength'] == 'strong'].shape[0]
print(f"Number of strong passwords after relaxing rules: {strongCount}")

moderateCount = dataset[dataset['strength'] == 'moderate'].shape[0]
print(f"Number of moderate passwords after relaxing rules: {moderateCount}")

weakCount = dataset[dataset['strength'] == 'weak'].shape[0]
print(f"Number of weak passwords after relaxing rules: {weakCount}")

<h1>5. Saving the Finalized Dataset</h1>

# Extract password column from temp_dataset
tempDataset = pd.read_csv("Featured-Dataset.csv")
passwordColumn = tempDataset['password']

# Insert the password column at the beginning of the main DataFrame (df)
dataset.insert(0, 'password', passwordColumn)
dataset

dataset.to_csv("Final-Dataset.csv", index=False)

***
